Fixed: Permissions & arguments
